added blit support with toolbar
created BlitToolbar, adapted from NavigationToolbar2QT that forces the background of CRF to none when using the buttons so that the blit updates properly after zooming, etc

Still need to add support for other buttons like save. Also see if other buttons need to be removed. Methods to override can be found in def of NavigationToolbar2QT OR NavigationToolbar2